use common trait for no-normal-case test cases
