+++ tests/PatternBasedNormaliserTest.php
@@ -4,10 +4,12 @@ namespace webignition\HtmlValidationErrorNormaliser\Tests;
 
 use webignition\HtmlValidationErrorNormaliser\NormalisedError;
 use webignition\HtmlValidationErrorNormaliser\PatternBasedNormaliser;
+use webignition\HtmlValidationErrorNormaliser\Tests\DataProvider\NoNormalFormDataProviderTrait;
 use webignition\HtmlValidationErrorNormaliser\Tests\DataProvider\PatternBasedNormalisedFormDataProviderTrait;
 
 class PatternBasedNormaliserTest extends \PHPUnit_Framework_TestCase
 {
+    use NoNormalFormDataProviderTrait;
     use PatternBasedNormalisedFormDataProviderTrait;
 
     /**
@@ -33,27 +35,6 @@ class PatternBasedNormaliserTest extends \PHPUnit_Framework_TestCase
     public function testNormaliseHasNoNormalisedForm($htmlErrorString)
     {
         $this->assertNull($this->patternBasedNormaliser->normalise($htmlErrorString));
-    }
-
-    /**
-     * @return array
-     */
-    public function normaliseHasNoNormalisedFormDataProvider()
-    {
-        return [
-            'empty' => [
-                'htmlErrorString' => '',
-            ],
-            'no normalisation needed [0]' => [
-                'htmlErrorString' => 'unterminated comment: found end of entity inside comment',
-            ],
-            'no normalisation needed [1]' => [
-                'htmlErrorString' => 'literal is missing closing delimiter',
-            ],
-            'no normalisation needed [2]' => [
-                'htmlErrorString' => 'character data is not allowed here',
-            ],
-        ];
     }
 
     /**
